+++ ChangeLog.md
@@ -431,4 +431,5 @@
 ### 2026.2.11
 - 优化 WebScreenCapturer 组件方法 - 添加额外等待时间应对某些异常加载现象；
 - 优化 BotTaskScheduler 及相关定时指令 - 优化任务命名相关逻辑；
-- 新增 BotOperator 组件 - 可用于主动需求 Bot 对象的操作。
+- 新增 BotOperator 组件 - 可用于主动需求 Bot 对象的操作；
+- 优化 项目结构。