+++ ChangeLog.md
@@ -433,5 +433,5 @@
 - 优化 BotTaskScheduler 及相关定时指令 - 优化任务命名相关逻辑；
 - 新增 BotOperator 组件 - 可用于主动需求 Bot 对象的操作；
 - 优化 OneTimeAlarm 一次性闹钟指令 - 支持更多时间格式 优化输出；
-- 优化 所有 getHelpForAI 方法 - 移除不必要的限权信息；
+- 优化 所有 getHelpForAI 方法 - 移除不必要的限权信息和部分指令帮助；
 - 优化 项目结构。